+++ fractals/src/main/java/me/catmousedog/fractals/paneloperators/fractals/inversemandelbrot/InverseJulia.java
@@ -21,7 +21,7 @@ public class InverseJulia extends MouseFractal {
 
 		functions = new Function[] { new IterativeFunction(this), new NormalizedFunction(this),
 				new PotentialFunction(this), new EscapeAngleFunction(this), new BinaryFunction(this),
-				new LambertFunction(this), new TestFunction(this) };
+				new LambertFunction(this) };
 		function = functions[0];
 	}